weather information

+++ src/components/ReminderModal.vue
@@ -9,11 +9,11 @@
         <form @submit.prevent="saveReminder">
           <div class="row">
             <div class="col-md-6">
-              <label for="text">Remidnder</label>
+              <label for="text">Reminder</label>
               <input
                 class="form-control mb-3"
                 type="text"
-                placeholder="Recordario"
+                placeholder="Reminder"
                 v-model="localReminder.text"
                 maxlength="30"
                 required
@@ -26,7 +26,7 @@
               <input
                 class="form-control mb-3"
                 type="text"
-                placeholder="Bogota"
+                placeholder="City"
                 v-model="localReminder.city"
                 required
                 @input="validateColor"
@@ -55,9 +55,26 @@
             </div>
           </div>
         </form>
+
+        <div class="mt-4" v-if="localReminder.weather">
+          <h5
+            v-if="
+              localReminder.weather.description ||
+              localReminder.weather.city ||
+              localReminder.weather.temperature
+            "
+          >
+            Weather Information
+          </h5>
+          <span v-if="localReminder.weather.description">
+            Decription: {{ localReminder.weather.description }} <br />
+            City: {{ localReminder.weather.city }} <br />
+            Temperature: {{ localReminder.weather.temperature }} °C
+          </span>
+        </div>
       </div>
 
-      <div class="modal-footer d-flex justify-content-between">
+      <div class="modal-footer d-flex justify-content-between mt-4">
         <div>
           <button
             class="btn btn-danger px-4 py-2"
@@ -95,49 +112,80 @@ import { v4 as uuidv4 } from "uuid";
 
 export default {
   props: ["date", "reminder"],
-
   data() {
     return {
       localReminder: this.reminder ? { ...this.reminder } : this.newReminder(),
       isEditing: !!this.reminder,
     };
   },
-
   methods: {
     ...mapActions(["addReminder", "editReminder", "deleteReminder", "fetchWeather"]),
+
     async saveReminder() {
       const selectedDate = new Date(this.date);
       const currentDate = new Date();
 
-      // Verificar si la fecha seleccionada es anterior a la fecha actual
       if (selectedDate < currentDate) {
-        alert("No puedes crear recordatorios en días anteriores a hoy.");
+        this.$swal({
+          title: "Ups",
+          icon: "warning",
+          text: "You can't create reminders in past days.",
+        });
         return;
       }
 
-      const weather = await this.fetchWeather({
+      this.localReminder.weather = await this.fetchWeather({
         city: this.localReminder.city,
         date: this.date,
       });
 
-      this.localReminder.weather = weather;
-
       if (this.isEditing) {
         this.editReminder(this.localReminder);
+        this.$swal({
+          title: "Reminder Updated",
+          icon: "success",
+          text: "The reminder has been updated successfully!",
+        });
       } else {
         this.localReminder.id = uuidv4();
         this.localReminder.date = this.date;
         this.addReminder(this.localReminder);
+        this.$swal({
+          title: "Reminder Created",
+          icon: "success",
+          text: "The reminder has been created successfully!",
+        });
       }
 
       this.$emit("close");
-      this.saveRemindersToLocalStorage(); // Guardar en localStorage después de actualizar o guardar
+      this.saveRemindersToLocalStorage();
     },
 
     deleteCurrentReminder() {
-      this.deleteReminder(this.localReminder.id);
-      this.$emit("close");
-      this.saveRemindersToLocalStorage(); // Guardar en localStorage después de eliminar
+      this.$swal
+        .fire({
+          title: "Are you sure?",
+          text: "You won't be able to revert this!",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonText: "Yes, delete",
+          cancelButtonText: "Cancel",
+          cancelButtonColor: "#d33",
+          confirmButtonColor: "#198754",
+        })
+        .then(async (result) => {
+          if (result.isConfirmed) {
+            await this.deleteReminder(this.localReminder.id);
+            this.$swal({
+              title: "The reminder has been deleted!",
+              icon: "success",
+            });
+            this.$emit("close");
+            this.saveRemindersToLocalStorage();
+          } else {
+            this.$emit("close");
+          }
+        });
     },
 
     closeModal() {
@@ -157,7 +205,7 @@ export default {
         city: "",
         time: "",
         color: "#FFFFFF",
-        weather: "",
+        weather: {},
       };
     },
